Modularized the classes as well as functionality
